[PLAT-9520] Add e2e tests for manual spans (#34)
* test: :test_tube: add manual-spans test fixture and feature

* test: :test_tube: Add test that spans can be created (and ended) before .start is called

* test: :white_check_mark: remove useless test

* add chromium only test scenario

* update maze-runner to 7.20.1

* don't clean workspaces after build-packages

* update platform assertion

* skip cleanup on CI

* increase debugging on ci

* pass through ENV from buildkite

* add debugging

* pass CI to environment

* skip build cleanup in docker

* don't cleanup files when debugging

* don't SKIP_CLEANUP in dockerfile

* update skip fixture cleanup

* update dockerfile to run without maze-runner image

* docker build maze-runner image

* update node install in dockerfile

* don't debug on ci

* fix fxitures directory path

* only test userAgentData on localhost

* simplify chromium-only script

* change * to And for readability

* improve test readability

* wait for 4 traces in pre-start-spans scenario

+++ test/browser/features/support/build-packages.rb
@@ -70,9 +70,11 @@ end
 
 # this at_exit is after the above commands on purpose - if they fail then we
 # DON'T want these cleanup commands to run so it's easier to debug
-at_exit do
-  Dir.chdir(FIXTURES_DIRECTORY) do
-    run("rm -rf node_modules")
-    run("npm run clean --workspaces")
+unless ENV.key?("SKIP_FIXTURE_CLEANUP") || ENV.key?("DEBUG")
+  at_exit do
+    Dir.chdir(FIXTURES_DIRECTORY) do
+      run("rm -rf node_modules")
+      run("npm run clean --workspaces")
+    end
   end
 end
